Transfer most of app.py into views.py
